--update namefiles, refactor functions
