+++ app.py
@@ -1,10 +1,11 @@
-from Classes.OllamaChat import OllamaChat
+from models.ollama_chat import OllamaChat
 import logging
 
 # Configuration du logging
-logging.basicConfig(level=logging.INFO)
-logger = logging.getLogger(__name__)
-logging.getLogger("httpx").setLevel(logging.WARNING)
+# logging.basicConfig(level=logging.INFO)
+# logger = logging.getLogger(__name__)
+# Désactiver les logs httpx
+# logging.getLogger("httpx").setLevel(logging.WARNING)
 
 def main():
     chat = OllamaChat(system_prompt='Tu es un assistant français sympathique et intelligent. Réponds de manière naturelle et concise.')
