+++ rag.py
@@ -1,22 +1,11 @@
-from Classes.OllamaChat import OllamaChat
-import PyPDF2
+from models.ollama_chat import OllamaChat
+from functions.extract_text_from_pdf import extract_text_from_pdf
 import logging
 
 # Configuration du logging
-logging.basicConfig(level=logging.INFO)
-logger = logging.getLogger(__name__)
-# Désactiver les logs httpx
-logging.getLogger("httpx").setLevel(logging.WARNING)
-
-
-# Fonction pour extraire le texte du fichier PDF
-def extract_text_from_pdf(file_path: str) -> str:
-    with open(file_path, 'rb') as file:
-        reader = PyPDF2.PdfReader(file)
-        text = ""
-        for page in reader.pages:
-            text += page.extract_text()
-    return text
+# logging.basicConfig(level=logging.INFO)
+# logger = logging.getLogger(__name__)
+# logging.getLogger("httpx").setLevel(logging.WARNING)
 
 def main():
     file_path = './pdf/temp.pdf'
